druid data source

+++ demo/src/main/resources/schema.sql
@@ -5,8 +5,8 @@ create table id_builder
 (
     id bigint auto_increment,
     name_space varchar(50) not null,
-    version bigint not null,
-    deleted bigint not null,
+    version int not null,
+    deleted int not null,
     gmt_create datetime not null,
     gmt_modified datetime not null,
     constraint id_builder_id_uk_index
@@ -25,8 +25,8 @@ create table invocation_store
     operator_name varchar(20) not null,
     operator_source varchar(50) not null,
     success tinyint not null,
-    version bigint not null,
-    deleted bigint not null,
+    version int not null,
+    deleted int not null,
     gmt_create datetime not null,
     gmt_modified datetime not null,
     constraint invocation_store_id_uk_index
@@ -46,8 +46,8 @@ create table item
     user_name   varchar(100) not null,
     amount bigint not null,
     store_code varchar(100),
-    version bigint not null,
-    deleted bigint not null,
+    version int not null,
+    deleted int not null,
     gmt_create datetime not null,
     gmt_modified datetime not null,
     constraint item_item_id_uk_index
@@ -64,8 +64,8 @@ create table inventory
     item_id bigint not null,
     amount bigint not null,
     store_code varchar(100),
-    version bigint not null,
-    deleted bigint not null,
+    version int not null,
+    deleted int not null,
     gmt_create datetime not null,
     gmt_modified datetime not null,
     constraint inventory_item_id_uk_index
@@ -84,8 +84,8 @@ create table message_store
     trigger_id    bigint       not null,
     class_name    varchar(200) null,
     invocation_id bigint       not null,
-    version bigint not null,
-    deleted bigint not null,
+    version int not null,
+    deleted int not null,
     gmt_create datetime not null,
     gmt_modified datetime not null,
     constraint message_store_id_uk_index
